Remove unneeded evaluate file
The functions are only used within the regression test. So move them into the regression test file and delete evaluate.py
