+++ src/kbmod/__init__.py
@@ -9,9 +9,9 @@ from . import kbmodpy
 from . import run_search
 from . import image_info
 from . import kbmod_info
-from . import evaluate
 from . import analysis_utils
 from . import jointfit_functions
+from . import result_list
 
 # lazy import analysis to arrest
 # loading large libraries in them
